forecasting/requirements.txt
pandas
numpy
pmdarima
statsmodels
google-cloud-bigquery
matplotlib
scikit-learn
prophet
joblib